Force uniform dtype on smooth_mix_wan22_t2v_fp16 transformer
from_single_file may leave some parameters (e.g. bias) in bfloat16
even when torch_dtype=float16 is specified; calling .to(dtype) ensures
all parameters share the same dtype.

+++ smooth_mix_wan22_t2v_fp16/pytorch/loader.py
@@ -91,7 +91,7 @@ class ModelLoader(ForgeModel):
             config=CONFIG_REPO,
             subfolder="transformer",
             torch_dtype=dtype,
-        )
+        ).to(dtype)
         self._transformer.eval()
         return self._transformer
